requirements.txt
discord.py>=2.0.0
mega.py>=1.0.8
pyyaml>=6.0
requests>=2.28.0
python-ffmpeg>=1.0.0
pytest>=7.0.0
